requirements.txt
langchain_openai 
langchain_cohere 
azure.ai.inference 
langchain_azure_ai 
python-dotenv
pandas
google-genai
scikit-learn
krippendorff
together